requirements.txt
Flask==2.2.2
Flask-MySQLdb==1.0.1
idna==3.3
Jinja2==3.1.2
mysql-connector-python==8.0.30
mysql-connector-repackaged==0.3.1
mysqlclient==2.1.1
pipreqs==0.4.11
protobuf==3.20.1
python-nmap==0.7.1
pytest==7.1.3